Prevent deletion of wallet with balance; refactor messages.
Closes dennisreimann/btcpayserver#8

+++ BTCPayServer.Plugins.LNbank/Pages/Wallets/Delete.cshtml.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Threading.Tasks;
 using BTCPayServer.Abstractions.Constants;
 using BTCPayServer.Client;
@@ -41,8 +42,15 @@ public class DeleteModel : BasePageModel
         });
 
         if (Wallet == null) return NotFound();
+        
+        if (Wallet.Balance > 0)
+        {
+            TempData[WellKnownTempData.ErrorMessage] = "This wallet still has a balance.";
+            return Page();
+        }
 
         await WalletService.RemoveWallet(Wallet);
+        TempData[WellKnownTempData.SuccessMessage] = "Wallet removed.";
 
         return RedirectToPage("./Index");
     }
